fixing g++ warnings in 400 and 401

+++ 401/777D.cpp
@@ -10,7 +10,7 @@ void solve() {
     }
     for (int i = n-2; i >=0; i--) {
         if (a[i].compare(a[i+1]) <= 0) { continue; }
-        int k = min(a[i].size(),a[i+1].size());
+        int k = min( (int) a[i].size(), (int) a[i+1].size());
         a[i].resize(k);
         for (int j = 0; j <k; j++) {
             if (a[i][j] > a[i+1][j]) { a[i].resize(j); break; }
